chore(p2p_network/client): move parsing to a separate file
